# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Lobsense IoT V2 API Documentation",
        "description": "Dokumentasi resmi API sistem monitoring, telemetri sensor tambak, pengontrolan aerator/feeder, inferensi AI, dan manajemen perangkat Lobsense V2.",
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Lobsense API Local Server"
        }
    ],
    "paths": {
...
        "/api/v2/operators": {
            "get": {
                "tags": [
                    "Operators"
                ],
                "summary": "Daftar Semua Operator",
                "description": "Mengambil seluruh data operator lapangan yang terdaftar.",
                "operationId": "e5cd01e017f789d7fe9c48925c5fd0dc",
                "responses": {
                    "200": {
                        "description": "Daftar operator berhasil diambil"
                    },
                    "401": {
                        "description": "Tidak terautentikasi"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Operators"
                ],
                "summary": "Tambah Operator Baru",
                "description": "Mendaftarkan operator lapangan baru ke dalam sistem.",
                "operationId": "1a186a04aa3fbe0af514b880e7faba21",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "full_name",
                                    "phone_number",
                                    "address"
                                ],
                                "properties": {
                                    "full_name": {
                                        "type": "string",
                                        "example": "Budi Setiawan"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "081234567890"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Jl. Pantai Indah No. 12, Mataram"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Operator berhasil ditambahkan"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v2/operators/{id}": {
            "get": {
                "tags": [
                    "Operators"
                ],
                "summary": "Detail Operator",
                "description": "Mengambil detail data satu operator berdasarkan ID.",
                "operationId": "a5a6c8ba7a55c7571dbf6d6c64df1b7f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail operator berhasil diambil"
                    },
                    "404": {
                        "description": "Operator tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Operators"
                ],
                "summary": "Update Operator",
                "description": "Memperbarui data operator lapangan yang sudah ada.",
                "operationId": "3e980c32dfb29526817816ba5d9bb58f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "full_name": {
                                        "type": "string",
                                        "example": "Budi Setiawan"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "081234567890"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Jl. Pantai Indah No. 15, Mataram"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operator berhasil diperbarui"
                    },
                    "404": {
                        "description": "Operator tidak ditemukan"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Operators"
                ],
                "summary": "Hapus Operator",
                "description": "Menghapus data operator berdasarkan ID.",
                "operationId": "df8e019959328cc6ed3c0033dcea5b4a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operator berhasil dihapus"
                    },
                    "404": {
                        "description": "Operator tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v2/provinces": {
            "get": {
                "tags": [
                    "Regions"
                ],
                "summary": "Daftar Provinsi",
                "description": "Mengambil seluruh data provinsi Indonesia.",
                "operationId": "5b37fb21dfe2e242613fab3d89746bf7",
                "responses": {
                    "200": {
                        "description": "Daftar provinsi berhasil diambil"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v2/cities": {
            "get": {
                "tags": [
                    "Regions"
                ],
                "summary": "Daftar Kota/Kabupaten",
                "description": "Mengambil data kota/kabupaten dengan filter opsional berdasarkan provinsi.",
                "operationId": "f6b2a763d446f09d787a1cacb3418953",
                "parameters": [
                    {
                        "name": "province_code",
                        "in": "query",
                        "description": "Kode provinsi untuk filter",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "52"
                        }
                    },
                    {
                        "name": "province_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar kota berhasil diambil"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v2/districts": {
            "get": {
                "tags": [
                    "Regions"
                ],
                "summary": "Daftar Kecamatan",
                "description": "Mengambil data kecamatan dengan filter opsional berdasarkan kode kota.",
                "operationId": "92ebd2c1f8b269c7f304185dc9b4b8c4",
                "parameters": [
                    {
                        "name": "city_code",
                        "in": "query",
                        "description": "Kode kota untuk filter",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "5201"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar kecamatan berhasil diambil"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}